include/sensors: import public pre-processor definitions types and to uorb.h
